# user-service/src/main/java/com/spring/user_service/controller/ProfileController.java
package com.spring.user_service.controller;

import com.spring.user_service.dto.profile.request.ProfilePostRequest;
import com.spring.user_service.dto.profile.response.ProfileGetResponse;
import com.spring.user_service.mapper.ProfileMapper;
import com.spring.user_service.service.ProfileService;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import jakarta.validation.Valid;
import lombok.AllArgsConstructor;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.Pageable;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.Set;
import java.util.stream.Collectors;

@RestController
@RequestMapping("v1/profiles")
@AllArgsConstructor
@SecurityRequirement(name = "basicAuth")
public class ProfileController {

    private final ProfileService profileService;
    private final ProfileMapper mapper;

    @GetMapping
    public ResponseEntity<Set<ProfileGetResponse>> findAllProfiles() {
        return ResponseEntity.ok(profileService.findAll().stream()
                .map(mapper::toProfileGetResponse).collect(Collectors.toSet()));
    }

    @GetMapping("/paginated")
    public ResponseEntity<Page<ProfileGetResponse>> findAllProfilesPaginated(Pageable pageable) {
        return ResponseEntity.ok(profileService.findAllPageable(pageable).map(mapper::toProfileGetResponse));
    }

    @GetMapping("/{id}")
    public ResponseEntity<ProfileGetResponse> findProfileById(@PathVariable Long id) {
        return ResponseEntity.ok(mapper.toProfileGetResponse(profileService.findById(id)));
    }

    @GetMapping("/by-name")
    public ResponseEntity<ProfileGetResponse> findProfileByName(@RequestParam String name) {
        return ResponseEntity.ok(mapper.toProfileGetResponse(profileService.findByName(name)));
    }

    @PostMapping
    public ResponseEntity<Void> saveProfile(@Valid @RequestBody ProfilePostRequest profilePostRequest) {
        var profile = mapper.toProfile(profilePostRequest);
        profileService.save(profile);
        return ResponseEntity.status(HttpStatus.CREATED).build();
    }

}